Refactor of DEC-939 for readability
* Moved some special function code to individual modules.
* Created smaller react components for the lists and document links.
