+++ build-conf/webpack-dev.config.js
@@ -46,7 +46,7 @@ module.exports = [
       new ExtractTextPlugin("resources/styles.css"),
       new webpack.DefinePlugin({
         "process.env": {
-          NODE_ENV: JSON.stringify("production")
+          NODE_ENV: JSON.stringify("development")
         }
       })
     ],
